Moved build ind vals to hdf utils
