+++ docs/auto_examples/dev_tutorials/plot_tutorial_02_writing_to_h5.py
@@ -85,6 +85,7 @@ import os
 from warnings import warn
 
 # Package for downloading online files:
+import pycroscopy.core.io.hdf_utils
 import pycroscopy.core.io.translator
 import pycroscopy.core.io.write_utils
 
@@ -280,9 +281,9 @@ px.plot_utils.plot_cluster_results_together(np.reshape(labels, (num_rows, num_co
 #
 # In this case, `centroids` has `k` positions all in one dimension. Thus the matrix needs to be reshaped to `k x 1`
 
-ds_labels_spec_inds, ds_labels_spec_vals = pycroscopy.core.io.write_utils.build_ind_val_dsets([1], labels=['Label'])
-ds_cluster_inds, ds_cluster_vals = pycroscopy.core.io.write_utils.build_ind_val_dsets([centroids.shape[0]], is_spectral=False,
-                                                                                      labels=['Cluster'])
+ds_labels_spec_inds, ds_labels_spec_vals = pycroscopy.core.io.hdf_utils.build_ind_val_dsets([1], labels=['Label'])
+ds_cluster_inds, ds_cluster_vals = pycroscopy.core.io.hdf_utils.build_ind_val_dsets([centroids.shape[0]], is_spectral=False,
+                                                                                    labels=['Cluster'])
 labels_mat = np.uint32(labels.reshape([-1, 1]))
 
 # Rename the datasets
