+++ pycroscopy/io/translators/gmode_iv.py
@@ -14,7 +14,7 @@ import h5py
 import numpy as np  # For array operations
 
 from ...core.io.translator import Translator, generate_dummy_main_parms
-from ...core.io.write_utils import build_ind_val_dsets
+from pycroscopy.core.io.hdf_utils import build_ind_val_dsets
 from ...core.io.hdf_utils import get_h5_obj_refs, link_h5_objects_as_attrs
 from ...core.io.hdf_writer import HDFwriter  # Now the translator is responsible for writing the data.
 from ...core.io.virtual_data import VirtualGroup, VirtualDataset  # building blocks for defining heirarchical storage in the H5 file
