+++ pycroscopy/io/translators/gmode_tune.py
@@ -17,7 +17,8 @@ from scipy.io.matlab import loadmat  # To load parameters stored in Matlab .mat 
 from .df_utils.be_utils import parmsToDict
 from .gmode_line import GLineTranslator
 from ...core.io.translator import generate_dummy_main_parms
-from ...core.io.write_utils import build_ind_val_dsets, VALUES_DTYPE
+from ...core.io.write_utils import VALUES_DTYPE
+from pycroscopy.core.io.hdf_utils import build_ind_val_dsets
 from ...core.io.hdf_utils import get_h5_obj_refs, link_h5_objects_as_attrs
 from ...core.io.hdf_writer import HDFwriter
 from ...core.io.virtual_data import VirtualGroup, VirtualDataset
